🎨Transfer pagination logic in usePagination custom hook
