+++ src/modules/views/Employees/EmployeesGallery/EmployeesGallery.jsx
@@ -5,45 +5,19 @@ import {Paginator} from '../../../components/Paginator/Paginator.jsx'
 import {PaginationLimiter} from '../../../components/PaginationLimiter/PaginationLimiter'
 import arrowLeft from '../../../../assets/icons/arrow/chevron_left.svg'
 import arrowRight from '../../../../assets/icons/arrow/chevron_right.svg'
+import {usePagination} from '../../../../hooks/usePagination.jsx'
 
 export const EmployeesGallery = () => {
-	const [page, setPage] = useState(0)
-	const [currentPage, setCurrentPage] = useState(page + 1)
+	const [page, currentPage, {setPrev, setNext, setPage}] = usePagination()
 	const [limit, setLimit] = useState(12)
 	const {data, isLoading, isFetched, isError} = useGetAllEmployees(page, limit, {enabled: true})
 	const numberOfPages = Math.ceil(data?.employeesLength / limit)
 	const firstPage = page === 0
 	const lastPage = page === numberOfPages - 1
 	
-	const handlePageNumber = (page) => {
-		setPage(page - 1)
-		setCurrentPage(page)
-	}
-	
-	const handleNextPage = () => {
-		if (page === numberOfPages - 1) {
-			setPage(numberOfPages - 1)
-			setCurrentPage(numberOfPages)
-		} else {
-			setPage(page + 1)
-			setCurrentPage(currentPage + 1)
-		}
-	}
-	
-	const handlePrevPage = () => {
-		if (page === -1) {
-			setPage(0)
-			setCurrentPage(1)
-		} else {
-			setPage(page - 1)
-			setCurrentPage(currentPage - 1)
-		}
-	}
-	
-	
-	const nextBtn = <button disabled={lastPage} className={lastPage ? 'icon icon-disabled' : 'icon'} onClick={handleNextPage}>
+	const nextBtn = <button disabled={lastPage} className={lastPage ? 'icon icon-disabled' : 'icon'} onClick={() => setNext(numberOfPages)}>
 		<img src={arrowRight} alt='next page'/></button>
-	const prevBtn = <button disabled={firstPage} className={firstPage ? 'icon icon-disabled' : 'icon'} onClick={handlePrevPage}>
+	const prevBtn = <button disabled={firstPage} className={firstPage ? 'icon icon-disabled' : 'icon'} onClick={setPrev}>
 		<img src={arrowLeft} alt='previous page'/></button>
 	
 	return isLoading ? <div>LOADING</div> : isError ? <div>ERROR</div> : (
@@ -68,7 +42,7 @@ export const EmployeesGallery = () => {
 				{numberOfPages > 1
 					? <div className='emp-gallery__pagination--nav'>
 						{prevBtn}
-						{<Paginator totalOfPages={numberOfPages} setPage={handlePageNumber} currentPage={currentPage}/>}
+						{<Paginator totalOfPages={numberOfPages} setPage={setPage} currentPage={currentPage}/>}
 						{nextBtn}
 					</div>
 					: <div className='no-pagination'></div>
